update.lessons

+++ Codes.in.C++/lessons/soma.natural.c
@@ -1,5 +1,5 @@
 #include<stdio.h>
-int function(int a,int b)
+int function(double a,double b)
 {   a--;
     if( a <= b)
         return b;
@@ -7,7 +7,7 @@ int function(int a,int b)
         return (a + function( a,b ));
 }
 int main(){
-    int a, b;
-    scanf("%d %d",&a, &b);
-      printf("%d\n",( b + function( b,a )));
+    double a, b;
+    scanf("%lf %lf",&a, &b);
+      printf("%.0lf\n",( b + function( b,a )));
 return 0;}
